pyproject.toml
[tool.black]
line-length = 90
target-version = ['py37']

[tool.poetry]
name = "bater"
version = "0.0.0"
description = "Battery manager."
license = "MIT"
authors = ["myslak71 <myslak@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.9"
attrs = "^19.3.0"
psutil = "^5.6.5"

[tool.poetry.dev-dependencies]
black = "^19.3.0"
pytest = "^3.0"
pytest-cov = "^2.4"
